Fix setting property of undefined in settings
The settings.js is executed even when we are not on the settings page.
This results in a setting property of undefined error when
document.getElementById('format').value is called.

That is because settings.js was present in content_scripts section.
The content_scripts section is used for scripts that you want to inject
into web pages (the pages that the user visits). However we only need
it on the settings page, so it only needs to be referenced in the
settings.html file.

See https://developer.chrome.com/docs/extensions/develop/concepts/content-scripts